Expose API to check whether the Siddhi App is stateful or not #1373
This expose queries, partitions, windows, triggers, and tables to check whether they are stateful or not.
