+++ findbugs-exclude.xml
@@ -97,7 +97,7 @@
         <Bug pattern="VO_VOLATILE_INCREMENT"/>
     </Match>
     <Match>
-        <Class name="io.siddhi.core.partition.PartitionRuntime"/>
+        <Class name="io.siddhi.core.partition.PartitionRuntimeImpl"/>
         <Bug pattern="RV_RETURN_VALUE_OF_PUTIFABSENT_IGNORED"/>
     </Match>
     <Match>
