+++ modules/siddhi-core/src/main/java/io/siddhi/core/stream/output/sink/Sink.java
@@ -29,6 +29,7 @@ import io.siddhi.core.util.SiddhiConstants;
 import io.siddhi.core.util.StringUtil;
 import io.siddhi.core.util.config.ConfigReader;
 import io.siddhi.core.util.parser.helper.QueryParserHelper;
+import io.siddhi.core.util.snapshot.state.EmptyStateHolder;
 import io.siddhi.core.util.snapshot.state.State;
 import io.siddhi.core.util.snapshot.state.StateFactory;
 import io.siddhi.core.util.snapshot.state.StateHolder;
@@ -344,5 +345,9 @@ public abstract class Sink<S extends State> implements SinkListener {
         LOG,
         WAIT,
         STREAM
+    }
+
+    public boolean isStateful() {
+        return stateHolder != null && !(stateHolder instanceof EmptyStateHolder);
     }
 }
