+++ modules/siddhi-core/src/main/java/io/siddhi/core/table/record/AbstractRecordTable.java
@@ -425,4 +425,9 @@ public abstract class AbstractRecordTable extends Table {
             return siddhiQueryContext;
         }
     }
+
+    @Override
+    public boolean isStateful() {
+        return false;
+    }
 }
